chore: separate functionalities into separate files to distiinguish migration checks from migrations creation
